requirements.txt
flask>=3.0.0
flask-login>=0.6.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
